# packages/api/swagger.yaml
openapi: 3.0.0
info:
  title: PolyScope API
  description: |
    PolyScope REST API - Analytics platform for Polymarket prediction markets.

    ## Features
    - Market discovery and search
    - Portfolio tracking
    - Position management
    - Real-time price feeds
    - Trading signals and news

    ## Authentication
    All endpoints except `/health` require an API key in the `X-API-Key` header.

    ## Rate Limiting
    - IP-based: 100 requests per minute
    - API key-based: 1000 requests per minute

    ## Response Format
    All responses follow a consistent structure:
    ```json
    {
      "success": true,
      "data": { ... },
      "meta": {
        "requestId": "uuid",
        "timestamp": "2026-02-03T14:00:00Z"
      }
    }
    ```
  version: 2.0.0
  contact:
    name: PolyScope Team
    url: https://github.com/yourusername/polyscope
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api.polyscope.example.com/api/v1
    description: Production
security:
  - ApiKeyAuth: []
paths:
...
  /markets:
    get:
      summary: List markets
      description: Get a list of markets with optional filtering
      parameters:
        - name: limit
          in: query
          description: Maximum number of markets to return (max 100)
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          description: Number of markets to skip
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: active
          in: query
          description: Filter by active status
          schema:
            type: boolean
            default: true
        - name: closed
          in: query
          description: Include closed markets
          schema:
            type: boolean
            default: false
        - name: category
          in: query
          description: Filter by category
          schema:
            type: string
            enum: [crypto, politics, sports, tech, macro, entertainment, science]
      responses:
        '200':
          description: List of markets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketsResponse'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /markets/search:
    get:
      summary: Search markets
      description: Search markets by keywords
      parameters:
        - name: q
          in: query
          description: Search query (min 2 characters)
          required: true
          schema:
            type: string
            minLength: 2
        - name: active
          in: query
          description: Filter by active status
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /markets/{slug}:
    get:
      summary: Get market details
      description: Get detailed information about a specific market
      parameters:
        - name: slug
          in: path
          required: true
          description: Market slug or identifier
          schema:
            type: string
      responses:
        '200':
          description: Market details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
...
  /news:
    get:
      summary: Get news
      description: Get latest news with trading signals
      parameters:
        - name: limit
          in: query
          description: Number of articles to return
          schema:
            type: integer
            default: 10
            maximum: 50
        - name: category
          in: query
          description: Filter by category
          schema:
            type: string
            enum: [crypto, politics, sports, tech, macro, entertainment, science]
        - name: minConfidence
          in: query
          description: Minimum confidence threshold (0-1)
          schema:
            type: number
            minimum: 0
            maximum: 1
        - name: minImpact
          in: query
          description: Minimum impact level
          schema:
            type: string
            enum: [low, medium, high]
            default: low
      responses:
        '200':
          description: News articles with signals
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsResponse'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /news/signals:
    get:
      summary: Get aggregated signals
      description: Get aggregated trading signals from news
      parameters:
        - name: market
          in: query
          description: Filter by specific market
          schema:
            type: string
        - name: timeframe
          in: query
          description: Timeframe for signals
          schema:
            type: string
            default: 24h
      responses:
        '200':
          description: Aggregated signals
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignalsResponse'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: BAD_REQUEST
              message: Invalid parameters
            meta:
              requestId: abc-123
              timestamp: "2026-02-03T14:00:00Z"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: Resource not found
            meta:
              requestId: abc-123
              timestamp: "2026-02-03T14:00:00Z"
    RateLimitExceeded:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retry
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: RATE_LIMIT_EXCEEDED
              message: Too many requests
              details:
                retryAfter: 60
            meta:
              requestId: abc-123
              timestamp: "2026-02-03T14:00:00Z"
  schemas:
    BaseResponse:
      type: object
      required:
        - success
        - meta
      properties:
        success:
          type: boolean
        meta:
          type: object
          properties:
            requestId:
              type: string
              format: uuid
            timestamp:
              type: string
              format: date-time
            cached:
              type: boolean
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            error:
              type: object
              required:
                - code
                - message
              properties:
                code:
                  type: string
                message:
                  type: string
                details:
                  type: object
...
    Market:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        question:
          type: string
        category:
          type: string
        prices:
          type: array
          items:
            type: number
        volume:
          type: object
          properties:
            total:
              type: number
            '24h':
              type: number
        liquidity:
          type: number
        status:
          type: object
          properties:
            active:
              type: boolean
            closed:
              type: boolean
            acceptingOrders:
              type: boolean
        endDate:
          type: string
          format: date-time
    MarketsResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                markets:
                  type: array
                  items:
                    $ref: '#/components/schemas/Market'
                pagination:
                  type: object
                  properties:
                    limit:
                      type: integer
                    offset:
                      type: integer
                    total:
                      type: integer
                    hasMore:
                      type: boolean
    SearchResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                results:
                  type: array
                  items:
                    $ref: '#/components/schemas/Market'
                query:
                  type: string
                matched:
                  type: boolean
                suggestions:
                  type: array
                  items:
                    type: object
    MarketDetailResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                id:
                  type: integer
                question:
                  type: string
                slug:
                  type: string
                prices:
                  type: object
                volume:
                  type: object
                liquidity:
                  type: object
                status:
                  type: object
                resolution:
                  type: object
...
    NewsResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                summary:
                  type: object
                articles:
                  type: array
                insights:
                  type: array
    SignalsResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                aggregate:
                  type: object
                  properties:
                    sentiment:
                      type: string
                    confidence:
                      type: string
                    signalCount:
                      type: integer
                topMarkets:
                  type: array